Revert "Reenable name search, add defensive code to convertKHPResourceAttributes"
