+++ hrm-form-definitions/form-definitions/zw/v1/ReferenceData.json
@@ -159,10 +159,6 @@
         {
           "value": "Umguza",
           "label": "Umguza"
-        },
-        {
-          "value": "Victoria Falls",
-          "label": "Victoria Falls"
         }
       ],
       "MTS": [
@@ -185,10 +181,6 @@
         {
           "value": "Beitbrigde",
           "label": "Beitbrigde"
-        },
-        {
-          "value": "Plumtree",
-          "label": "Plumtree"
         }
       ],
       "MSE": [
