+++ hrm-form-definitions/form-definitions/zw/v1/tabbedForms/IssueCategorizationTab.json
@@ -240,9 +240,6 @@
         },
         {
           "label": "Maintenance"
-        },
-        {
-          "label": "Inheritance"
         }
       ]
     },
