+++ plugin-hrm-form/package.json
@@ -55,7 +55,6 @@
     "react-keyboard-event-handler": "^1.5.4",
     "react-redux": "^7.2.6",
     "react-router-dom": "^5.3.0",
-    "react-viewport-list": "^7.1.0",
     "redux": "^4.1.2",
     "redux-promise-middleware-actions": "^3.1.0",
     "rollbar": "^2.25.0",
