--- plugin-hrm-form/src/hooks/useTraceUpdate.ts
@@ -1,39 +1,0 @@
-/**
- * Copyright (C) 2021-2023 Technology Matters
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see https://www.gnu.org/licenses/.
- */
-
-/*
- * Lifted from https://stackoverflow.com/a/51082563
- * Dont leave calls to it in code long term, it's just for debugging
- */
-
-import { useEffect, useRef } from 'react';
-
-// eslint-disable-next-line import/no-unused-modules
-export default function useTraceUpdate(props, name: string) {
-  const prev = useRef(props);
-  useEffect(() => {
-    const changedProps = Object.entries(props).reduce((ps, [k, v]) => {
-      if (prev.current[k] !== v) {
-        ps[k] = [prev.current[k], v];
-      }
-      return ps;
-    }, {});
-    if (Object.keys(changedProps).length > 0) {
-      console.log(`[${name}]`, 'Changed props:', changedProps);
-    }
-    prev.current = props;
-  });
-}
