+++ plugin-hrm-form/src/utils/setUpComponents.tsx
@@ -45,8 +45,6 @@ import { Container } from '../styles/queuesStatus';
 import { FeatureFlags, isInMyBehalfITask } from '../types/types';
 import { colors } from '../channels/colors';
 import { getHrmConfig } from '../hrmConfig';
-import AseloMessageInput from '../components/AseloMessageInput';
-import AseloMessageList from '../components/AseloMessageList';
 
 type SetupObject = ReturnType<typeof getHrmConfig>;
 /**
@@ -357,14 +355,6 @@ export const removeActionsIfTransferring = () => {
   Flex.ParticipantCanvas.Content.remove('actions', {
     if: props => hasNoControlAndIsWarm(props.task) && props.participant.participantType === 'worker',
   });
-};
-
-/**
- *
- */
-export const replaceTwilioMessageInput = () => {
-  Flex.MessageInputV2.Content.replace(<AseloMessageInput key="textarea" />, { sortOrder: -1 });
-  Flex.MessageList.Content.replace(<AseloMessageList key="list" />);
 };
 
 /**
