+++ twilio-iac/terraform-modules/channels/flow-templates/language-mt/with-chatbot.tftpl
@@ -481,7 +481,7 @@ ${
         "service": "{{trigger.message.InstanceSid}}",
         "channel": "{{trigger.message.ChannelSid}}",
         "from": "{{flow.channel.address}}",
-        "body": "Do you accept our Terms and Conditions? https://kellimni.com/chat-terms/ \nPlease answer with a number:\n1. Yes, I accept the terms and conditions.\n2. No, I don't accept the terms and conditions.",
+        "body": "Do you accept our Terms and Conditions?  https://kellimni.com/about/terms-conditions/ \nPlease answer with a number:\n1. Yes, I accept the terms and conditions.\n2. No, I don't accept the terms and conditions.",
         "timeout": "3600"
       }
     },
@@ -600,7 +600,7 @@ ${
         "service": "{{trigger.message.InstanceSid}}",
         "channel": "{{trigger.message.ChannelSid}}",
         "from": "{{flow.channel.address}}",
-        "body": "Inti taċċetta t-Termini u l-Kundizzjonijiet tagħna? https://kellimni.com/chat-terms/ \nJekk jogħġbok wieġeb billi tagħżel numru:\n1. Iva, naċċetta t-Termini u l-Kundizzjonijiet. \n2. Le, ma naċċettax t-Termini u l-Kundizzjonijiet.",
+        "body": "Inti taċċetta t-Termini u l-Kundizzjonijiet tagħna? https://kellimni.com/about/terms-conditions/ \nJekk jogħġbok wieġeb billi tagħżel numru:\n1. Iva, naċċetta t-Termini u l-Kundizzjonijiet. \n2. Le, ma naċċettax t-Termini u l-Kundizzjonijiet.",
         "timeout": "3600"
       }
     },
@@ -719,7 +719,7 @@ ${
         "service": "{{trigger.message.InstanceSid}}",
         "channel": "{{trigger.message.ChannelSid}}",
         "from": "{{flow.channel.address}}",
-        "body": "Ви приймаєте наші умови? https://kellimni.com/chat-terms/ \n Будь ласка, дайте відповідь цифрою:\n1. Так, я приймаю умови. \n2. Ні, я не приймаю умови.",
+        "body": "Ви приймаєте наші умови? https://kellimni.com/about/terms-conditions/ \n Будь ласка, дайте відповідь цифрою:\n1. Так, я приймаю умови. \n2. Ні, я не приймаю умови.",
         "timeout": "3600"
       }
     },
